+++ metaflow/plugins/docker/docker_cli.py
@@ -2,7 +2,7 @@ import os
 from metaflow._vendor import click
 from metaflow.cli import echo_always as echo
 
-from .bakery import read_metafile, BAKERY_METAFILE
+from .fast_bakery import read_metafile, BAKERY_METAFILE
 
 
 @click.group()
